Rename folder structure as convention
