+++ src/routes.py
@@ -3,11 +3,11 @@ from flask_restful import Api
 
 from src.resources.device.device_plural import DevicePlural
 from src.resources.device.device_singular import DeviceSingularByName, DeviceSingularByUUID
-from src.resources.mappings.mapping import LPGBPMappingResourceList, LPGBPMappingResourceByLoRaPointUUID, \
+from src.resources.mapping.mapping import LPGBPMappingResourceList, LPGBPMappingResourceByLoRaPointUUID, \
     LPGBPMappingResourceByGenericPointUUID, LPGBPMappingResourceByBACnetPointUUID
 from src.resources.network.network import SerialDriver
 from src.resources.ping import Ping
-from src.resources.points.points import PointsSingularByName, PointsSingularByUUID, PointsPlural
+from src.resources.point.point import PointsSingularByName, PointsSingularByUUID, PointsPlural
 
 bp_lora = Blueprint('lora', __name__, url_prefix='/api/lora')
 api_lora = Api(bp_lora)
@@ -19,7 +19,7 @@ api_lora.add_resource(DeviceSingularByName, '/devices/name/<string:value>')
 api_lora.add_resource(PointsSingularByUUID, '/points/uuid/<string:value>')
 api_lora.add_resource(PointsSingularByName, '/points/name/<string:value>')
 
-# lora to generic/bacnet points mapping
+# lora to generic/bacnet points mappings
 bp_lp_gbp_mapping = Blueprint('lp_gbp_mappings', __name__, url_prefix='/api/lp_gbp/mappings')
 api_lp_gbp_mapping = Api(bp_lp_gbp_mapping)
 api_lp_gbp_mapping.add_resource(LPGBPMappingResourceList, '')
